requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
gspread==6.0.0
oauth2client==4.1.3
aiohttp==3.9.1